Refactor Galfus Runner and Runtime for Improved Module Resolution and Entry Point Handling
- Removed the galfus-vm dependency from galfus-runner and added it to galfus-runtime.
- Introduced a new WorkspaceResolver to handle module source resolution and reading.
- Updated ModuleLoader to utilize WorkspaceResolver for loading modules and resolving imports.
- Enhanced error handling for entry functions in the Runtime, ensuring proper validation of exported functions.
- Modified the run_project function to support custom entry points defined in the workspace configuration.
- Updated tests to validate the new entry point handling and ensure exported functions are required.
- Refactored IO handling in the VM to utilize target capabilities instead of direct IO handlers.
- Improved overall code organization and readability across multiple files.

+++ crates/galfus-ir/src/lower/expression.rs
@@ -480,7 +480,7 @@ impl<'a, 'b> FnEmitter<'a, 'b> {
                 let prim = match constant {
                     MirConstant::Null => PrimitiveType::Null,
                     MirConstant::Bool(_) => PrimitiveType::Bool,
-                    MirConstant::Int(_) => PrimitiveType::Int64,
+                    MirConstant::Int(_) => PrimitiveType::Int32,
                     MirConstant::Float(_) => PrimitiveType::Float64,
                     MirConstant::String(_) => {
                         // Find String type in type table
@@ -490,10 +490,10 @@ impl<'a, 'b> FnEmitter<'a, 'b> {
                                 table.kind(ty_id),
                                 Some(TypeKind::Primitive(PrimitiveType::Uint8))
                             ) {
-                                // Fallback to Int64 if not found
+                                // Fallback to Int32 if not found
                             }
                         }
-                        PrimitiveType::Int64
+                        PrimitiveType::Int32
                     }
                 };
                 for i in 0..table.len() {
